web/requirements.txt
gradio==4.4.1
gradio_client==0.7.0
imageio==2.9.0
json5==0.9.6
jsonpath==0.82
keras==2.9.0
keras-bert==0.89.0
keras-embed-sim==0.10.0
keras-layer-normalization==0.16.0
keras-multi-head==0.29.0
keras-pos-embd==0.13.0
keras-position-wise-feed-forward==0.8.0
Keras-Preprocessing==1.1.2
keras-self-attention==0.51.0
numpy==1.22.2
numpydoc==1.2
openai==0.28.1
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
Pillow==9.0.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-lsp-black==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
python-multipart==0.0.6
python-slugify==5.0.2
python-snappy==0.6.0
python-utils==3.4.5
scikit-build==0.16.4
scikit-image==0.19.2
scikit-learn==1.1.3
sklearn-features==0.0.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
torch==2.1.1
torchvision==0.16.1
pymysql==1.1.0
pandas==2.2.0
Bottleneck==1.3.6